split model package object
