+++ src/main/scala/nl/knaw/dans/easy/multideposit/model/package.scala
@@ -23,9 +23,6 @@ package object model {
   type DatasetId = String
   type DepositorId = String
 
-  type DatasetRow = Map[MultiDepositKey, String]
-  type DatasetRows = Seq[DatasetRow]
-
   // inspired by http://stackoverflow.com/questions/28223692/what-is-the-optimal-way-not-using-scalaz-to-type-require-a-non-empty-list
   type NonEmptyList[A] = ::[A]
 
@@ -39,9 +36,5 @@ package object model {
       if (list.isEmpty) List(default)
       else list
     }
-  }
-
-  implicit class DatasetRowFind(val row: DatasetRow) extends AnyVal {
-    def find(name: MultiDepositKey): Option[String] = row.get(name).filterNot(_.isBlank)
   }
 }
