+++ src/main/scala/nl/knaw/dans/easy/multideposit/parser/package.scala
@@ -1,0 +1,13 @@
+package nl.knaw.dans.easy.multideposit
+
+import nl.knaw.dans.easy.multideposit.model.MultiDepositKey
+
+package object parser {
+
+  type DatasetRow = Map[MultiDepositKey, String]
+  type DatasetRows = Seq[DatasetRow]
+
+  implicit class DatasetRowFind(val row: DatasetRow) extends AnyVal {
+    def find(name: MultiDepositKey): Option[String] = row.get(name).filterNot(_.isBlank)
+  }
+}
