+++ src/test/scala/nl/knaw/dans/easy/multideposit/model/ModelPackageSpec.scala
@@ -17,7 +17,7 @@ package nl.knaw.dans.easy.multideposit.model
 
 import nl.knaw.dans.easy.multideposit.UnitSpec
 
-class ParserPackageSpec extends UnitSpec {
+class ModelPackageSpec extends UnitSpec {
 
   "NonEmptyList transform" should "succeed when the input list is not empty" in {
     listToNEL(List(1, 2, 3)) shouldBe ::(1, ::(2, ::(3, Nil)))
@@ -34,25 +34,5 @@ class ParserPackageSpec extends UnitSpec {
 
   it should "return a list with the default value if the input list is empty" in {
     List.empty.defaultIfEmpty(-1) shouldBe List(-1)
-  }
-
-  "find" should "return the value corresponding to the key in the row" in {
-    val row = Map("foo" -> "abc", "bar" -> "def")
-    row.find("foo").value shouldBe "abc"
-  }
-
-  it should "return None if the key is not present in the row" in {
-    val row = Map("foo" -> "abc", "bar" -> "def")
-    row.find("baz") shouldBe empty
-  }
-
-  it should "return None if the value corresponding to the key in the row is empty" in {
-    val row = Map("foo" -> "abc", "bar" -> "")
-    row.find("bar") shouldBe empty
-  }
-
-  it should "return None if the value corresponding to the key in the row is blank" in {
-    val row = Map("foo" -> "abc", "bar" -> "  \t  ")
-    row.find("bar") shouldBe empty
   }
 }
